+++ acms-react-app/src/Services/admin.service.js
@@ -5,6 +5,7 @@ const API_BASE_URL = process.env.REACT_APP_BACKEND_URL || "http://localhost:8080
 
 const API_CUSTOMERS_URL = `${API_BASE_URL}/api/v1/customers`;
 const API_APPOINTMENTS_URL = `${API_BASE_URL}/api/v1/appointments`;
+const API_INVOICES_URL = `${API_BASE_URL}/api/v1/invoices`
 
 class adminService{
 
@@ -79,6 +80,13 @@ class adminService{
 
     deleteAppointment(id){
         return axios.delete(API_APPOINTMENTS_URL + `/${id}`, { headers: authHeader() });
+    }
+
+
+    // invoices
+
+    getAllInvoices(){
+        return axios.get(API_INVOICES_URL, { headers: authHeader() });
     }
 
 }
